% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microservices-package.R
\docType{package}
\name{microservices-package}
\alias{microservices}
\alias{microservices-package}
\title{microservices: Breakdown a Monolithic Application to a Suite of Services}
\description{
'Microservice' architectural style is an approach to developing a 
    single application as a suite of small services, each running in its own 
    process and communicating with lightweight mechanisms, often an 'HTTP' 
    resource 'API'. These services are built around business capabilities and 
    independently deployable by fully automated deployment machinery. There 
    is a bare minimum of centralized management of these services, which may 
    be written in different programming languages and use different data storage 
    technologies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidylab/microservices}
  \item Report bugs at \url{https://github.com/tidylab/microservices/issues}
}

}
\author{
\strong{Maintainer}: Harel Lustiger \email{tidylab@gmail.com} (\href{https://orcid.org/0000-0003-2953-9598}{ORCID})

Other contributors:
\itemize{
  \item Tidylab [copyright holder, funder]
}

}
\keyword{internal}
